Moves Handler._reset_db_connection() call to get_base_paths_blocking()
This change ensures that the connection is reset inside the same thread that is
used to perform ORM calls.

This fixes a bug where the /pulp/content/ page would return a 500 error after the
connection the database is closed due to a network problem or a database restart.

Required PR: https://github.com/pulp/pulp-smash/pull/1261

fixes: #9515
https://pulp.plan.io/issues/9515